refactor: Source code layout for ConfigAuditReports
