+++ src/configauditreports/details.scss
